+++ packages/xchain-aggregator/src/types.ts
@@ -106,6 +106,7 @@ type QuoteSwapParams = {
   destinationAddress?: string // The destination address for the swap
   height?: number // The block height for the swap
   toleranceBps?: number // The tolerance basis points for the swap
+  liquidityToleranceBps?: number // The pool swing tolerance points (optional)
   streamingQuantity?: number // The streaming quantity for the swap
   streamingInterval?: number // The streaming interval for the swap
 }