+++ packages/xchain-doge/src/utils.ts
@@ -3,8 +3,8 @@
  */
 import { Network } from '@xchainjs/xchain-client'
 import { Address } from '@xchainjs/xchain-util'
+import { toBitcoinJS } from '@xchainjs/xchain-utxo'
 import * as Dogecoin from 'bitcoinjs-lib' // Importing bitcoinjs-lib for Dogecoin operations
-import coininfo from 'coininfo' // Importing coininfo for cryptocurrency information retrieval
 
 /**
  * Constant values representing transaction sizes and lengths.
@@ -45,18 +45,11 @@ export function arrayAverage(array: number[]): number {
 export const dogeNetwork = (network: Network): Dogecoin.networks.Network => {
   switch (network) {
     case Network.Mainnet:
-      return coininfo.dogecoin.main.toBitcoinJS()
+      return toBitcoinJS('dogecoin', 'main') as Dogecoin.networks.Network
     case Network.Stagenet:
-      return coininfo.dogecoin.main.toBitcoinJS()
+      return toBitcoinJS('dogecoin', 'main') as Dogecoin.networks.Network
     case Network.Testnet: {
-      // Latest coininfo on NPM doesn't contain dogetest config information
-      const bip32 = {
-        private: 0x04358394,
-        public: 0x043587cf,
-      }
-      const test = coininfo.dogecoin.test
-      test.versions.bip32 = bip32
-      return test.toBitcoinJS()
+      return toBitcoinJS('dogecoin', 'test') as Dogecoin.networks.Network
     }
   }
 }