+++ packages/xchain-mayachain/src/client.ts
@@ -1,7 +1,13 @@
 import { StdFee } from '@cosmjs/amino'
 import { Bip39, EnglishMnemonic, Secp256k1, Slip10, Slip10Curve } from '@cosmjs/crypto'
 import { fromBase64, toBase64 } from '@cosmjs/encoding'
-import { DecodedTxRaw, DirectSecp256k1HdWallet, EncodeObject, TxBodyEncodeObject, decodeTxRaw } from '@cosmjs/proto-signing'
+import {
+  DecodedTxRaw,
+  DirectSecp256k1HdWallet,
+  EncodeObject,
+  TxBodyEncodeObject,
+  decodeTxRaw,
+} from '@cosmjs/proto-signing'
 import { DeliverTxResponse, SigningStargateClient } from '@cosmjs/stargate'
 import { AssetInfo, Network, PreparedTx, TxHash } from '@xchainjs/xchain-client'
 import {