+++ examples/do-swap/CHANGELOG.md
@@ -1,5 +1,39 @@
 # xchainjs-do-swap
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [a8988c5]
+- Updated dependencies [a8988c5]
+  - @xchainjs/xchain-thorchain@3.0.4
+  - @xchainjs/xchain-bsc@2.0.2
+  - @xchainjs/xchain-ethereum@2.0.2
+  - @xchainjs/xchain-thorchain-amm@3.0.6
+  - @xchainjs/xchain-wallet@2.0.5
+
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies [561e2a4]
+  - @xchainjs/xchain-thorchain-query@2.0.1
+  - @xchainjs/xchain-thorchain-amm@3.0.5
+  - @xchainjs/xchain-thornode@1.0.1
+
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [89585ee]
+  - @xchainjs/xchain-evm-providers@2.0.1
+  - @xchainjs/xchain-ethereum@2.0.1
+  - @xchainjs/xchain-avax@2.0.1
+  - @xchainjs/xchain-bsc@2.0.1
+  - @xchainjs/xchain-evm@2.0.1
+  - @xchainjs/xchain-thorchain-amm@3.0.4
+  - @xchainjs/xchain-wallet@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes