+++ examples/prepareTx/CHANGELOG.md
@@ -1,5 +1,19 @@
 # xchainjs-prepare-tx
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [a8988c5]
+  - @xchainjs/xchain-ethereum@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [89585ee]
+  - @xchainjs/xchain-ethereum@2.0.1
+
 ## 2.0.0
 
 ### Major Changes