+++ packages/xchain-mayachain-amm/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 4.0.5
+
+### Patch Changes
+
+- Updated dependencies [a8988c5]
+- Updated dependencies [a8988c5]
+  - @xchainjs/xchain-thorchain@3.0.4
+  - @xchainjs/xchain-ethereum@2.0.2
+  - @xchainjs/xchain-wallet@2.0.5
+
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [89585ee]
+  - @xchainjs/xchain-arbitrum@2.0.1
+  - @xchainjs/xchain-ethereum@2.0.1
+  - @xchainjs/xchain-evm@2.0.1
+  - @xchainjs/xchain-wallet@2.0.4
+
 ## 4.0.3
 
 ### Patch Changes